HISTFILE is gone (V)('M')(V)
Long-term history suggestions aren't all that useful in my
experience-
Normally, when I create a new shell,
I just want to repeat the set of commands that
shell was created for.

If I have some long command I've needed to do a million times,
I'll have written a short alias for it.